Move CellState to a separate file
