+++ src/core/structure/cell/cellState.ts
@@ -1,0 +1,7 @@
+export enum CellState {
+  OK,
+  INVALID_EXPRESSION,
+  INVALID_REFERENCE,
+  CIRCULAR_REFERENCE,
+  INVALID_FORMAT,
+}
