+++ src/core/structure/sheet.ts
@@ -1,11 +1,12 @@
 import { CellKey, ColumnKey, RowKey } from "./key/keyTypes.ts";
-import Cell, { CellState } from "./cell/cell.ts";
+import Cell from "./cell/cell.ts";
 import Column from "./group/column.ts";
 import Row from "./group/row.ts";
 import { CellInfo, PositionInfo } from "./sheet.types.ts";
 import ExpressionHandler from "../evaluation/expressionHandler.ts";
 import CellStyle from "./cellStyle.ts";
 import CellGroup from "./group/cellGroup.ts";
+import { CellState } from "./cell/cellState.ts";
 
 export default class Sheet {
   defaultStyle: any;
