+++ src/core/structure/sheet.types.ts
@@ -1,5 +1,6 @@
 import { ColumnKey, RowKey } from "./key/keyTypes.ts";
-import { CellState } from "./cell/cell.ts";
+
+import { CellState } from "./cell/cellState.ts";
 
 export type PositionInfo = {
   columnKey?: ColumnKey;
